Changed alot
Added all mechanics yet some pieces are not working. Putting the blocks into the bins does not trigger a message. Trying to unpause after pausing does not allow the character to continue moving or pause again. Some of the roots of the objects are  off and do not allow for rotation or proper movement. The collisions have all been fixed. It is possible to finish the level before completing the tasks yet doing so does not trigger the game over screen.